Added snippetFailedFunc
When there's a problem rendering a snippet, each of the
functions in the list are called.  By default, the
list is populated with a Log.warn, but you
can add other functions (probably in Boot).

The most valuable function to add would be a function
that indicates a failure during testing.



git-svn-id: https://liftweb.googlecode.com/svn/trunk/liftweb@478 a1e4567e-5129-0410-9a71-e7582cf7e72e